requirements.txt
asgiref==3.5.2
certifi==2021.10.8
dj-rest-auth==2.2.4
Django==4.0.5
django-cors-headers==3.13.0
django-filter==21.1
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
wincertstore==0.2
